correct naming normalization problems
In generating type definitions for various service APIs, I noticed the
normalization of names for fields and structs had a few errors. For
example, in the RDS API, the following are all field names:

* DBInstanceIdentifier
* DbInstanceIdentifier
* DBInstanceId
* DbiResourceId

Pay attention to the slight case differences and the use of the word
"Identifier" and "Id" or "ID"...

This commit adds fixes for incorrect normalization that was, for
example, outputting "DbInstanceIDentifier" or "dBiResourceID".