remove exception classes from type defs
Exceptions should not be output as type definitions since they will not
be included in any CRD's Spec or Status struct. This patch takes
advantage of a feature added to the `aws-api-tool schema` command which
outputs an `x-aws-api-exception: true` annotation into the Swagger
schema for shapes with an exception attribute.