begin plumbing findSDKXXX methods in res manager
This patch adds a simplistic implementation of the
AWSResourceManager.ReadOne() interface method for the Bookstore API
example book resource manager. We follow a similar strategy to the
AppMesh v1beta2 implementation, farming out most real work to a
findSDKBook() method that communicates with the AWS service API via the
aws-sdk-go.

Besides the above findSDKBook() code, the remainder of this patch is
simply a move of the services/example/pkg/sdk package into
services/example/sdk. This directory/package does NOT appear in the real
service controllers since the import will be the
aws-sdk-go/services/$SERVICE package.