CUD implementation for bookstore example res mgr (#46)
* simple implementation of bookstore res mgr Create

Adds a simple implementation of the example Book resource manager's
Create() method.

* simplistic implementation of book res mgr Update

Adds a very simple implementation to the Book resource manager's
Update() method. Adds note to the AWSResourceManager.Update() interface
method docstring to indicate that it is NOT expected for the resource
manager's Update() method to determine whether the desired and latest
observed state of the resource differ. This is the responsibility of the
reconciler, which lives at a higher layer in the ACK runtime.

* simple implementation for book res mgr Delete

Adds a simple Delete implementation for the example Book resource
manager.