place concrete ACK types into apis/core/v1alpha
`pkg/types` is for only interfaces, not concrete types. This patch moves
the concrete types `AWSAccountID` and `AWSResourceName` into a new
package `apis/core/v1alpha1` and adds some new annotation constants and
a new concrete `ResourceMetadata` struct to that package.

The reason it is important to separate the interfaces from the concrete
types is because the concrete types are used from within the CRDs
themselves and therefore should be "versioned" -- i.e. the package
should be a versioned package like `apis/core/v1alpha1` so that we can
evolve the CRD definitions in an orderly fashion over time.

The interfaces that are in `pkg/types`, on the other hand, define the
programming interface that the controller/managers structs must
implement.

The AWSResource interface gets a new `Identifiers()` method in this
patch, which returns a object that implements a new
`AWSResourceIdentifiers` interface. This `AWSResourceIdentifiers`
interface contains methods that allow the individual service API
resource constructs to return ARNs, AWS Account IDs, and other
identifiers, in a consistent fashion.