Add Identifiers().ARN() method and integrate logs
Add the ARN() method to the `AWSResourceIdentifiers` interface and add
implementation for the example bookstore resource. Also adds a new
`AdoptedResourceNotFound` error and code path in the `reconciler.sync`
method to handle the case for when the backend resource isn't found but
the user specified an ARN in the CR's annotations (indicating that the
ACK service controller should "adopt" the resource).