Make sure that rmCache is initialized
Varun noticed in testing that the ManagerFor() method of the
bookResourceManagerFactory struct was throwing a nil pointer exception
due to the rmCache field not being properly initialized to a valid map.
This patch fixes that problem by ensuring that the resource manager
factory registered in init() has a properly initialized rmCache.