correct lookup of referred-to object schemas
When determining the Go type for schemas of type "object" or "array of
object", we were searching for the referred-to schema using the wrong
name. For the case of schemas with a type of "object", we were passing
the name of the property, camel-cased. In the case of the "array of
object" type schemas, we were using the name of the property instead of
the name of the schema being referred to.

What this meant was that, for example, the ECR API's ImageScanFindings
object's "findings" property, which looks like this:

```yaml
    ImageScanFindings:
      properties:
        findingSeverityCounts:
          $ref: '#/components/schemas/FindingSeverityCounts'
        findings:
          $ref: '#/components/schemas/ImageScanFindingList'
        imageScanCompletedAt:
          $ref: '#/components/schemas/ScanTimestamp'
        vulnerabilitySourceUpdatedAt:
          $ref: '#/components/schemas/VulnerabilitySourceUpdateTimestamp'
      type: object
    ImageScanFindingList:
      items:
        properties:
          attributes:
            $ref: '#/components/schemas/AttributeList'
          description:
            $ref: '#/components/schemas/FindingDescription'
          name:
            $ref: '#/components/schemas/FindingName'
          severity:
            $ref: '#/components/schemas/FindingSeverity'
          uri:
            $ref: '#/components/schemas/Url'
        type: object
      type: array
    ImageScanFinding:
      properties:
        attributes:
          $ref: '#/components/schemas/AttributeList'
        description:
          $ref: '#/components/schemas/FindingDescription'
        name:
          $ref: '#/components/schemas/FindingName'
        severity:
          $ref: '#/components/schemas/FindingSeverity'
        uri:
          $ref: '#/components/schemas/Url'
      type: object
```

we were looking for a schema named "Findings" instead of
"ImageScanFindings".

Fixes Issue #64