collate all Operations for a CRD from schema
Following patches rely on being able to determine the payload of an
operation in order to generate the appropriate SDK-specific objects in
the AWSResourceManager templated code. This patch creates a map, keyed
by operation type (e.g. OpTypeCreate, OpTypeGet, etc) and resource name
(the Kind of the CRD) to an *openapi3.Operation struct that the
templates will use to query for payload information.