fix pluralized referred-to object types in arrays
The RDS API's DBCluster resource has a schema that looks like this:

```yaml
    DBCluster:
      properties:
        ... lots of other fields ...
        DBClusterOptionGroupMemberships:
          $ref: '#/components/schemas/DBClusterOptionGroupMemberships'
    DBClusterOptionGroupMemberships:
      items:
        properties:
          DBClusterOptionGroupName:
            $ref: '#/components/schemas/String'
          Status:
            $ref: '#/components/schemas/String'
        type: object
      type: array
```

Note that there is no "DBClusterOptionGroupMembership" struct.

Many AWS APIs needlessly (due to XML legacy) wrap arrays of some
singular object in structs that are nothing more than a struct with a
single field that is an array of the singular object type. There is code
in ACK that looks for this situation and "unwraps" those needless
wrapper structs and makes the end Go type in the CRD's Spec struct a
slice of pointer to the singular object type.

So, instead of generating this:

```go
type MyField struct {
    Foo string
}
type MyFieldList struct {
    Foos []*Foo
}
type MyResourceSpec struct {
    Foos *MyFieldList
}
```

we generate the (much simpler and readable):

```go
type MyField struct {
    Foo string
}
type MyResourceSpec struct {
    Foos []*Foo
}
```

The code that looks at the above array-of-object-type fields and deduces
that the Go type for the `DBCluster.Spec.DBClusterOptionGroupName` field
was attempting to look for a schema with a singular version of the
"DBClusterOptionGroupMemberships" name. But there isn't one.

This patch fixes that situation and falls back to looking up the
pluralized version of the original array type name.

Fixes Issue #76 however we still need to figure out how to handle cases
where an "anonymous array element struct" is found. This occurs in, for
example, again in the RDS API, the following schema:

```yaml
    OptionGroupOptionVersionsList:
      items:
        properties:
          IsDefault:
            $ref: '#/components/schemas/Boolean'
          Version:
            $ref: '#/components/schemas/String'
        type: object
      type: array
    OptionGroupOptionsList:
      items:
        properties:
... lots of other fields ....
          OptionGroupOptionVersions:
            $ref: '#/components/schemas/OptionGroupOptionVersionsList'
        type: object
      type: array
```

There is no "OptionGroupOptionVersion" schema, nor is there an
"OptionGroupOptionVersions" schema. Instead, there is only an
"OptionGroupOptionVersionsList" schema that is itself an array type
containing an unnamed (anonymous) object element type containing two
properties "IsDefault" and "Version". :(

(Also note that there is actually an "OptionVersion" named
schema/struct! For some reason, the RDS API authors have the
OptionVersionsList struct contain elements of type "OptionVersion", but
the "OptionGroupOptionVersionsList" struct is different for some unknown
reason...)

Since there is no way for us to handle this in the current code, I just
return an error back up from the deduceArrayOfObjectsType() function and
let the Go type for the field just be "AnonymousArrayObjectType".
Clearly that will need to be addressed in the future but for now, it
gets the api/typedef generation working again for APIs like RDS.