add SDKMapper and normalization of Status attrs
Each API acts differently and the input/output shapes for individual API
calls follow different, inconsistent conventions. We need a way to
standardize many of these inconsistencies.

This patch adds a new SDKMapper object to the pkg/schema package. This
struct's purpose is to contain instructions to the code generator for
generating the SDK calls in the resource manager.

The first standardization that the SDKMapper helps with is identifying
the fields in an Operation's output shape that contain the primary
resource's ARN information. In constructing the model.CRD now, in
pkg/schema's `Helper.GetCRDs()` method, we now remove from the list of
the CRD's Status fields any field that has the same name as a field in
the corresponding Spec struct. In addition, if we find a field called
"arn" or "${RESOURCE}arn" in the output shape, we indicate to the
SDKMapper that this field stores the primary object's ARN information
and remove it from the Status fields.

See new tests of the SNS Topic, EC2 LaunchTemplate and ECR Repository's
CRD generation and lots of comments in the test assertions for more
information on how these normalizations occur.