support standard package types using import map
Adds support to the services/$SERVICE/apis/$VERSION/types.go and
individual CRD .go file generation for non-builtin types from standard
packages.

For example, in the ECR CreateRepository output shape, there is a
CreateTime field that is of type `*time.Time`.

This patch adds a map of package paths to add for importing at the top
of the types.go and individual CRD .go files. This gets the generated
code to build properly without throwing undefined import errors.