correct unused imports and variables in sdk gen
When the Create, Update, and Delete operations are called, the code
generated had the form:

```go
    resp, err := rm.sdkapi.CreateResourceWithContext(ctx, input)
...
    // Add fields to CR's Status from output..
    ko := r.ko.DeepCopyObject()
    ko.Status.SomeField = resp.SomeField
```

Which is fine, however if there are no StatusFields in the CR (which
happens when there are no fields in the Create output struct that are
not already in the CreateInput struct and thus already in the CR.Spec
struct) then the `resp` variable is unused and you'll run into failures
when building the controller.

This patch fixes the above issue and a similar problem with the
importing of the awserr and ACK's pkg/errors when there is no ReadOne
SDK API call and thus no use of those packages.