DRY up the Go type and pkg alias handling
There was a bunch of code that was repeated in the code generator that
output a CRD's Go code that constructed Input and Output shapes. This
patch places all that repeated code into a helper function
`model.CRD.replacePkgName()` and addresses an issue I ran into: we only
want to replace the primary aws-sdk-go/services API package, not
standard imports like "time".