fix struct name conflicts and Go type cleaning
There were issues with enum and type names that conflicted with the
CRD's Spec and Status types. For instance, in the RDS API there is a
DBProxyStatus enum type that conflicts with the DBProxy CRD's Status
field type, which is `type DBProxyStatus struct`. We identify these
conflicting names and suffix the generated types with "_SDK".

Secondly, I was not generating the proper Go types properly. I need to
use the normalized Camel-cased name of the struct or list element type
instead of the name returned by
aws-sdk-go/private/model/api.Shape.GoTypeWithPkgName().

Fixes Issue #89