introduce framework for handling "special" APIs
We needed a way of informing the code generator that certain APIs use
"special" handling of certain fields for resources by using an
"Attributes" `map[string]*string` that contains actual, schema'd fields
for the resource.

An example of this type of special API is SNS API which has a
CreateTopic call that defines a Topic resource. The CreateTopic call has
an "Attributes" parameter that contains a set of key/value pairs for
fields such as "DeliveryPolicy" or "DisplayName".

We want to inform the code generator that these fields are contained in
the Attributes parameter and should be "unpacked" into the CRD
definition as top-level fields on the resource.

This patch introduces this functionality by creating a GeneratorConfig
struct that is represented as a YAML file that contains instructions to
the code generator on how to handle these kinds of
special/inconsistent/idiosyncratic API behaviours.

Issue #109