generate code for GetAttributes input/output (#116)
Adds generation for Go code that sets the Input shape for GetAttributes
operations or sets CRD Status fields from the Output shape of a
GetAttributes operation.

We add special handling for the primary resource ARN and the AWS Owner
Account ID attribute keys, if found in the Attributes map. For these
special cases, we route the returned Output shape attribute key to the
`Status.ACKResourceMetadata.ARN` and
`Status.ACKResourceMetadata.OwnerAccountID` fields respectively.

Issue #109