plumb reconciler into resource manager
We need a way of allowing the resource manager to get Kubernetes object
information for things like Kubernetes Secrets. However, I deliberately
did not want to pass the Kubernetes client objects into the resource
managers (since their job is to handle the SDK linkage, not to query
Kubernetes API).

However, the resource manager will need at least the ability to grab a
value for a SecretReference, and in order to facilitate this, I changed
the `AWSResourceManagerFactory.ManagerFor` interface method to pass in
an object that implements the `AWSResourceReconciler` interface. I then
added a stub `AWSResourceReconciler.SecretValueFromReference()`
interface method (for Alina to implement) into the runtime's Reconciler
concrete struct. The resource managers now have an `rr` field that
points to the Reconciler for the runtime and will allow the resource
managers to call utility methods like `SecretValueFromReference` and not
have to handle calls to the Kubernetes API directly themselves.