properly handle NotFoundException in reads
The templated sdk.go file was looking for the string "NotFoundException"
to identify whether a call to an AWS API was returning a 404 NotFound.
Unfortunately this isn't correct for most APIs, including Elasticache.
The exception codes in Elasticache look like "CacheClusterNotFound"
instead of the generic "NotFoundException". Trying to match on the
string "NotFound" doesn't work either, since calls to, for example, the
DescribeCacheClusters API call legitimately return 400 with a code of
"ServiceLinkedRoleNotFound" which is actually a permissions problem.

This patch adds a `CRD.NotFoundException()` method which looks through
the read-like API calls for Error codes and looks for 404 HTTP status
codes and writes the matched exception code string into the sdk.go
template instead of the hard-coded 'NotFoundException'.

In the process, I also fixed up a shadowing problem where some of the
sdk.go template code referred to the wrong error variable (err instead
of respErr).

Finally, a few cleanups to `scripts/kind-build-test.sh` allow properly
testing with AWS_REGION set and in environments where the default
output encoding for the `aws` CLI tool isn't `json`.