add ExceptionsConfig instructions for code gen
Adds a new ExceptionsConfig struct to the GeneratorConfig. This new
struct provides instructions to the code generator on how to identify
the Exception "Code" for a returned string error message for a
particular API resource's ReadOne/ReadMany/GetAttributes operations.

The `CRD.NotFoundException()` method has been changed to
`CRD.ExceptionCode()` and now accepts an HTTP status code that the
caller wants the string Exception code for a particular resource.

I've gone ahead and git add'd `generator.yaml` files for each of the
APIs where we've needed to give the code generator special instructions
into the `service/$ALIAS/generator.yaml` location. This will solve the
common problem that folks have been running into where they try
generating an API and run into errors like this:

```
error calling GoCodeGetAttributesSetOutput: called
GoCodeGetAttributesSetOutput for a resource `GetQueueAttributes` that
doesn't unpack attributes map
```

The above is due to the code generator recognizing an API uses the
GetXXXAttributes pattern but the code generator was not provided a
`generator.yaml` file instructing it how to handle the attribute maps.