address issues in ReadMany generated output code
When generating Go code that sets the Spec or Status fields of a CR from
the response Output shape of a List/ReadMany operation, we were
failing to use the "cleaned" names for Kubernetes API types in the
`CRD.goCodeVarEmptyConstructorK8sType()` method. While we were properly
looking up to see if the Shape had to be renamed (due to conflicts with
core objects like the CRDStatus or CRDSpec types), we were not cleaning
the type definition names for struct, map[string]*struct, or []*struct
types. This patch fixes that and adds unit tests of the RDS DBInstance
ReadMany output Go code generation to ensure properly cleaned type names
are used.

Issue #169

A second issue was identified when generating the RDS service
controller: there are actually API top-level resources that have a
Create operation but have no Read operation. Yes... RDS'
CreateDBInstanceReadReplica is an example of this. It creates a read
replica resource of an existing DB instance but there is no ReadOne,
ReadMany, or GetAttributes API calls for those read replicas.

This was leading to generated code throwing build errors about missing
return values from sdkFind() methods. The fix for this was to modify the
template for sdk.go to add an `{{ else }}` catchall that does:

```go
return nil, ackerr.NotImplemented
```

allowing the build to proceed past that point.