add ability to rename input fields
The S3 CreateBucket API Operation has an Input shape with a field called
"Bucket". This is actually the *name* of the Bucket to be created.
Unfortunately, the ListBuckets API Operation's Output shape (from which
we build the SDK linkage to find a particular object) calls this same
field "Name" (quite appropriately I might add, since that's what the
field represents... the bucket's name).

In order to handle mismatched field names like this, I have added a
`RenamesConfig` struct to the `GeneratorConfig` that holds instructions
on how to rename certain fields in the input and output shapes.

The S3 `generator.yaml` config has been modified with the following
rename instructions:

```yaml
resources:
  Bucket:
    renames:
      operations:
        CreateBucket:
          input_fields:
            Bucket: Name
```

which effectively renames the `BucketSpec.Bucket` field to
`BucketSpec.Name`.

Fixes the last build failure identified in #180.

Issue #180