add List operation match field filtering
Adds `ListOperationConfig` struct to the `GeneratorConfig`. This struct
allows us to instruct the code generator that a particular API's List
operation returns all resources of a particular type and that it will
need to add Go code to the `sdkFind()` method of the resource manager to
filter out results from the returned set of objects that do not match
the values of certain fields.

It's easier to talk about this with an example. Let's take the S3 Bucket
API. There is a ListBuckets API call. The Input shape for ListBuckets is
always empty. There is no ability to ask ListBuckets to only return
results for, say, a particular named Bucket. Instead, in order to use
the ListBuckets API call to find a *particular* named Bucket, you need
to ListBuckets and then loop over all the results, looking for a result
with a Name field matching the Bucket.Spec.Name that you are looking
for.

This patch changed the code generator to add Go code to the output shape
processing for the ListBuckets API that looks like this:

```go
if len(resp.Buckets) == 0 {
    return nil, ackerr.NotFound
}
found := false
for _, elem := range resp.Buckets {
    if elem.Name != nil {
        if ko.Spec.Name != nil {
            if *elem.Name != *ko.Spec.Name {
                continue
            }
        }
        r.ko.Spec.Name = elem.Name
    }
    found = true
}
if !found {
    return nil, ackerr.NotFound
}
```

The `generator.yaml` for S3 looks like this now:

```yaml
ignore:
  resource_names:
    - Object
    - MultipartUpload
  shape_names:
    # These shapes are structs with no members...
    - SSES3
resources:
  Bucket:
    renames:
      operations:
        CreateBucket:
          input_fields:
            Bucket: Name
        DeleteBucket:
          input_fields:
            Bucket: Name
    list_operation:
      match_fields:
        - Name
```

The `resources.Bucket.list_operation` element contains the instructions
telling the code generator to output special code to filter out results
that don't match the searched-for Bucket's Name.