SNS service controller codegen fixes
Quite a few problems with the generation of the SNS service controller.
When I wrote the smoke test for SNS, it identified issues with how the
call to GetTopicAttributes was being incorrectly made. The TopicArn
field in the GetTopicAttributesInput shape was not being set. So, I
added code to set that value to a simulated ARN (when the ARN doesn't
exist yet due to the resource not being created yet in the AWS API, we
can construct an ARN using the AWS account ID and region associated with
the resource manager).

This way, the GetTopicAttributes API call will return a 404 as
appropriate and the sdkFind() code paths will work properly.

I'm still, however, getting failures when attempting to run the smoke
tests. I'm getting the following error:

```
2020-08-17T01:22:53.772Z	ERROR	controller-runtime.controller	Reconciler error	{"controller": "topic", "request": "default/ack-test-smoke-sns", "error": "InvalidParameter: Invalid parameter: TopicArn Reason: A ${AWS_REGION} ARN must begin with arn:null, not arn:aws:sns:${AWS_REGION}:${AWS_ACCOUNT_ID}:ack-test-smoke-sns\n\tstatus code: 400, request id: 9b8aea74-c027-5dc4-bf0a-bdc71e0a7af2"}
```

and I'm lost as to how to interpret it. Hoping to get some help in the
morning from others...