ensure scripts/kind-build-test.sh has AWS Role ARN
The MissingRegion error we were seeing in test runs was actually due to
the AWS_ROLE_ARN environment variable not being set. This was due to me
running `./scripts/kind-build-test.sh -p -s ecr -r "$ROLE_ARN"` in a
terminal where I had not source/exported my IAM Role used for testing.
This resulted in the ECR service controller starting up, but a ""-valued
AWS_ROLE_ARN environment variable being placed into the Deployment env.
It was this that was "missing a region"... :(

This PR makes the `-r $ROLE_ARN` a required flag for
`scripts/kind-build-test.sh` to avoid problems like this in the future.