ensure ECR DescribeRepositories properly generated
There were two issues with the generated ECR service controller code.
The first issue had to do with the generated code not properly
understanding the NotFound exception code for repositories. The solution
to this was adding something to the generatorConfig for ECR to instruct
the generator which exception code to identity NotFound errors.

The second issue had to do with another generatorConfig instruction:
tell the code generator to output Go code to filter out Repository
results in the `DescribeRepositoriesOutput.Repositories` struct field
that did not have a match `RepositoryName` value as the CR's
`Spec.RepositoryName` field.

Adding the `services/ecr/generator.yaml` file and configuring it like so
fixed the above two problems:

```yaml
resources:
  Repository:
    exceptions:
      codes:
        404: RepositoryNotFoundException
    list_operation:
      match_fields:
        - RepositoryName
```

Finally, I updated both the unit test for ECR repository code generation
as well as modified the ECR smoke test to include creation of 3 separate
repositories and then delete those 3 repositories. Doing this triggers
the issue with DescribeRepositories returning multiple results and the
generated Go code not properly filtering out results by matching
RepositoryName.