update templates to correct gofmt issues
When generating the ECR service controller and getting ready to check it
into source control, I noticed a couple changes when I ran gofmt against
the generated code. Some of those changes were due to things like spaces
instead of TAB characters in the templates.

This fixes those up and runs `gofmt` against the generated code for a
service controller when we build it.

Issue #213