NotFound when any required field missing in shape
In #271, @vijtrip2 added methods that output Go code that checked if the
Input shape for ReadOne operations for a resource had required fields
and if so, checked that the CR's `Status.{Field}` had a non-nil value,
otherwise the `resourceManager.sdkFind()` method would return
`NotFound`, indicating to callers (in particular when trying to
determine if a newly-created resource already existed) that the backend
AWS resource had not been created yet.

This patch makes that work generic in two ways:

* adds support for when required fields in an Input shape are in the
  CR's `Spec` struct in addition to the `Status` struct.
* adds support for any Input shape, not just the ReadOne operation's
  Input shape.

For SQS, the GetQueueAttributes' Input shape has required fields
(QueueUrl) that, if not present in the CR's `Status` struct, indicate
the resource has not been created yet.

Issue #287