handle SetAttributes-based update operations
This patch adds to the code generator for the SDK linkage for APIs that
use the SetAttributes-flavor of updating a resource. Previously, the
`sdk.go` files for these APIs had empty `sdkUpdate` operations with a
big TODO that ended up breaking the service controller (quite
predictably).

While implementing this patch, I stumbled upon yet another unexplainable
inconsistency between two SetAttributes operations within the SNS
service API. The SNS `SetPlatformApplicationAttributes` API call accepts a
field in its Input shape called `Attributes` that is a map of key/value
pairs for attributes to set on the platform application. Unfortunately,
the similarly-named SNS `SetTopicAttributes` API call apparently does
*NOT* work the same way. Instead, there is a single `AttributeName` and
`AttributeValue` field in the Input shape and you need to call the
`SetTopicAttributes` API call once for each modified attribute. :(

In fact, the [official documentation][0] for the SNS SetTopicAttributes
API call says that the `AttributeName` field in the Input shape is
actually a "*map* of attributes with their corresponding values". But
that isn't actually the case...

So, the code in this patch for now just contains a TODO in the
`CRD.GoCodeSetAttributesSetInput()` method for SetAttributes APIs that
can only operate on a single attribute at a time. We will use the
CustomOperation functionality as a temporary workaround for these APIs
while we come up with a more permanent code-generated solution.

Issue #296

[0]: https://docs.aws.amazon.com/sns/latest/api/API_SetTopicAttributes.html