allow total override of Update op
Some APIs, like ECR, have no Update operation at all for some resources
(like Repository). Instead, there are a series of API calls like
`PutImageScanningConfiguration`, `PutImageTagMutability`, or
`SetRepositoryPolicy` that set individual attributes for a particular
resource.

This patch adds the ability to override the implementation of
`sdkUpdate` **entirely** to a custom method name of the resourceManager
struct, thereby allowing total control over the various update
operations. This fixes a bug where, for the ECR sdk.go file, the
sdkUpdate method was empty and returning `(nil, nil)` which was causing
segfaults in the service controller.

It also adds a simple Update code path test to the ECR smoke tests.