allow override of GetAttributes input shape
The SQS GetQueueAttributes API call is entirely different from the SNS
GetTopicAttributes call. Whereas the latter is called and returns the
Topic's attributes, the former has a required `[]*string` field on the
Input shape called `AttributeNames` that has to have a single "All"
element in it in order for the API call not to return an error about
malformed Input data :(

This patch introduces new structures in the generator configuration that
allows the code generator to write override values for the Input shape
in order to populate this properly.

Issue #205