add `ack-generate release` command
Adds a new `ack-generate release <SERVICE> <VERSION>` command that
generates a Helm chart for a specific service controller and release
version combination.

The Helm chart that is generated installs the Docker image for the
ACK controller for that service and release version. For example, if you
run:

```bash
ack-generate release s3 v0.0.1
```

the Helm chart generated will be placed in
`/services/s3/helm/Chart.yaml` and the image referred to from the
Deployment created in that chart will be
`amazon/aws-controllers-k8s:s3-v0.0.1`.

Also includes documentation on the release steps and a test of the Helm
installation process.

Issue #366