reworking service controller release roadmap
Customers and users are looking for a clear visual roadmap that shows
roughly when they should expect certain AWS services to get an ACK
controller into developer preview, beta and GA release statuses.

We will be re-shaping the projects/1 GH project to contain fine-grained
tasks and this PR reworks the services.md and releases.md documents to
turn the services.md doc into more of a clear tabular roadmap of service
controllers, with links to milestones (that can be date-targeted) for
individual service controllers.