update install docs and add helm package script
Updates the user installation docs to better highlight the way our
Docker images are tagged and how to properly install with Helm.

Adds a new `scripts/helm-package-controller.sh` script that runs the
`helm package` command against a service controller's generated Helm
chart to create the tarball package that is needed to place in the Helm
repository we'll be indexing in our gh-pages branch of the
aws-controllers-k8s source repository.