crossplane: generate hooks.go if not exists
The crossplane code generation was generating a `zz_hooks.go` file in
each resource subpackage. This was incorrect in two ways: first, the
file should not have been prefixed with `zz_`. Secondly, we should not
have been overwriting the hooks.go files if there was already one there,
since the hooks.go file can contain custom update code that developers
have modified as necessary.

This PR moves the generation of the `hooks.go` template out of the loop
over the zz_ target files for each CRD inside the crossplane TemplateSet
constructor and ensures that an existing hooks.go file is not
overwritten.

Also removes unncessary templates that crossplane doesn't use and
corrects a mistake in the hooks.go.tpl for the ReadMany code path's
isUpToDate() function.