ResourceConfig.NameField -> FieldConfig.IsName
The ResourceConfig.NameField allows a particular field in a CRD to be
marked as the "name field". If this configuration was set, the default
behaviour of looking for a field called "Name", "{Resource}Name" or
"{Resource}Id" was overridden.

This patch simply makes this "name_field" string attribute of the
ResourceConfig be instead a bool attribute of the FieldConfig, bringing
this configuration value in-line with the other FieldConfig boolean
attributes like "IsRequired", "IsPrintable" and "IsReadOnly".